complete optional task

+++ src/main/java/ru/javawebinar/topjava/web/MealServlet.java
@@ -22,53 +22,67 @@ public class MealServlet extends HttpServlet {
     private static final int MAX_CALORIES = 2000;
     private static final Logger log = getLogger(MealServlet.class);
     private final MealService mealService = new MealServiceImpl();
+    private static final String MEALS_JSP = "/meals.jsp";
+    private static final String EDITOR_JSP = "/editor.jsp";
 
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         log.debug("redirect to meals");
-
-        String action = request.getHttpServletMapping().getMatchValue();
+        String targetJsp = "";
+        String action = request.getParameter("action") == null ? "meals" : request.getParameter("action");
+        System.out.println(action);
         int id;
-        switch (action) {
-            case "delete":
-                id = Integer.parseInt(request.getParameter("id"));
-                log.debug("delete meal with id = " + id);
-                mealService.delete(id);
-                break;
-            case "update":
-                id = Integer.parseInt(request.getParameter("id"));
-                log.debug("update meal with id = " + id);
-                Meal meal = mealService.get(id);
-                request.setAttribute("myMeal", meal);
-                request.getRequestDispatcher("editor.jsp").forward(request, response);
-                break;
-            case "create":
-                log.debug("create meal");
-                request.getRequestDispatcher("editor.jsp").forward(request, response);
-                break;
+        if (action.equalsIgnoreCase("delete")) {
+            id = Integer.parseInt(request.getParameter("id"));
+            log.debug("delete meal with id =" + id);
+            mealService.delete(id);
+            response.sendRedirect("meals");
+            return;
+        } else if (action.equalsIgnoreCase("meals")) {
+            log.debug("show list of meals");
+            targetJsp = MEALS_JSP;
+            List<MealTo> mealTos = getTos();
+            request.setAttribute("mealTos", mealTos);
+        } else if (action.equalsIgnoreCase("update")) {
+            targetJsp = EDITOR_JSP;
+            id = Integer.parseInt(request.getParameter("id"));
+            log.debug("update meal with id =" + id);
+            Meal meal = mealService.get(id);
+            request.setAttribute("meal", meal);
+        } else if (action.equalsIgnoreCase("create")) {
+            log.debug("create meal");
+            targetJsp = EDITOR_JSP;
         }
-        List<Meal> meals = mealService.getAll();
-        List<MealTo> mealsTo = MealsUtil.filteredByStreams(meals, LocalTime.MIN,
-                LocalTime.MAX, MAX_CALORIES);
 
-        request.setAttribute("allMeals", mealsTo);
-        request.getRequestDispatcher("/meals.jsp").forward(request, response);
+        request.getRequestDispatcher(targetJsp).forward(request, response);
+    }
+
+    private List<MealTo> getTos() {
+        List<Meal> meals = mealService.getAll();
+        return MealsUtil.filteredByStreams(meals, LocalTime.MIN,
+                LocalTime.MAX, MAX_CALORIES);
     }
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        Meal meal = new Meal();
-        meal.setDateTime(LocalDateTime.parse(request.getParameter("dateTime")));
-        meal.setDescription(request.getParameter("description"));
-        meal.setCalories(Integer.parseInt(request.getParameter("calories")));
-
-        List<MealTo> mealsTo = MealsUtil.filteredByStreams(mealService.getAll(), LocalTime.MIN,
-                LocalTime.MAX, MAX_CALORIES);
-        request.setAttribute("allMeals", mealsTo);
-        request.getRequestDispatcher("/meals.jsp").forward(request, response);
+            throws IOException {
+        request.setCharacterEncoding("UTF-8");
+        Meal meal;
+        LocalDateTime localDateTime = LocalDateTime.parse(request.getParameter("dateTime"));
+        String description = request.getParameter("description");
+        int calories = Integer.parseInt(request.getParameter("calories"));
+        if (!request.getParameter("id").equals("")) {
+            int id = Integer.parseInt(request.getParameter("id"));
+            meal = new Meal(id, localDateTime, description, calories);
+        } else {
+            meal = new Meal();
+            meal.setDateTime(LocalDateTime.parse(request.getParameter("dateTime")));
+            meal.setDescription(request.getParameter("description"));
+            meal.setCalories(Integer.parseInt(request.getParameter("calories")));
+        }
+        mealService.save(meal);
+        response.sendRedirect("meals");
     }
-
 }
